reactor: move expandable row into separate component
